refactor: move Assist pipeline tools into their own module
tools_voice_assistant.py had grown to 1078 lines carrying two unrelated
concerns. ha_manage_pipeline and its helpers move verbatim into
tools_assist_pipeline.py as AssistPipelineTools; entity exposure keeps
the old module, so KNOWN_ASSISTANTS and every patch target in the
exposure tests stay where they are. The registry discovers the new
module by convention, so nothing is wired by hand.

Moving the pipeline half rather than the exposure half is what keeps
this small: it touches four files instead of the six the other
direction needs.
